ddb_bits: make shift overflow consistent by typing unity

+++ src/ddb_bits.h
@@ -7,7 +7,7 @@
 static uint32_t read_bits(const char *src, uint64_t offs, uint32_t bits)
 {
     uint64_t *src_w = (uint64_t*)&src[offs >> 3];
-    return (*src_w >> (offs & 7)) & ((1 << bits) - 1);
+    return (*src_w >> (offs & 7)) & ((((uint64_t)1) << bits) - 1);
 }
 
 static void write_bits(char *dst, uint64_t offs, uint32_t val)
